feat: ShowOneSlider 컴포넌트 생성
